refactor: move wallet errors into their dedicated file
